# docs/swagger.yaml
basePath: /
definitions:
...
  departments.BatchCreateDepartmentRequest:
    properties:
      departments:
        items:
          $ref: '#/definitions/departments.CreateDepartmentRequest'
        type: array
    type: object
  departments.BatchDeleteDepartmentRequest:
    properties:
      public_ids:
        example:
        - dept-abc123
        - dept-xyz789
        items:
          type: string
        type: array
    type: object
  departments.BatchOperationResponse:
    properties:
      failed_count:
        example: 0
        type: integer
      failed_ids:
        items:
          type: string
        type: array
      success_count:
        example: 5
        type: integer
    type: object
  departments.BatchUpdateDepartmentRequest:
    properties:
      updates:
        items:
          properties:
            description:
              type: object
            name:
              type: object
            parent_department_public_id:
              example: dept-parent123
              type: string
            public_id:
              example: dept-abc123
              type: string
          type: object
        type: array
    type: object
  departments.CreateDepartmentRequest:
    properties:
      description:
        type: object
      name:
        type: object
      parent_department_public_id:
        example: dept-parent123
        type: string
    type: object
  departments.DepartmentListResponseWrapper:
    properties:
      data:
        items:
          $ref: '#/definitions/departments.DepartmentResponse'
        type: array
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total_count:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  departments.DepartmentResponse:
    properties:
      description:
        type: object
      id:
        example: 1
        type: integer
      name:
        type: object
      parent_department_public_id:
        example: dept-parent123
        type: string
      public_id:
        example: dept-abc123
        type: string
    type: object
  departments.DepartmentTreeResponse:
    properties:
      children:
        items:
          $ref: '#/definitions/departments.DepartmentTreeResponse'
        type: array
      description:
        type: object
      id:
        example: 1
        type: integer
      name:
        type: object
      parent_department_public_id:
        example: dept-parent123
        type: string
      public_id:
        example: dept-abc123
        type: string
    type: object
  departments.ErrorResponse:
    properties:
      error:
        example: Bad Request
        type: string
      message:
        example: Invalid department name
        type: string
    type: object
  departments.MoveDepartmentRequest:
    properties:
      new_parent_department_public_id:
        example: dept-parent123
        type: string
    type: object
  departments.SearchDepartmentRequest:
    properties:
      description:
        example: tech
        type: string
      name:
        example: Engineering
        type: string
    type: object
  departments.SuccessResponse:
    properties:
      message:
        example: Department deleted successfully
        type: string
    type: object
  departments.UpdateDepartmentRequest:
    properties:
      description:
        type: object
      name:
        type: object
      parent_department_public_id:
        example: dept-parent123
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@knowledgecenter.io
    name: API Support
  description: Knowledge Center REST API Server
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  title: Knowledge Center API
  version: "1.0"
paths:
...
  /departments:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of departments
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/departments.DepartmentListResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List departments
      tags:
      - departments
    post:
      consumes:
      - application/json
      description: Creates a new department with JSONB name and description
      parameters:
      - description: Department data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/departments.CreateDepartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/departments.DepartmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new department
      tags:
      - departments
  /departments/{publicId}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes a department by public ID
      parameters:
      - description: Department Public ID
        in: path
        name: publicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/departments.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete department
      tags:
      - departments
    get:
      consumes:
      - application/json
      description: Retrieves a department by its public ID
      parameters:
      - description: Department Public ID
        in: path
        name: publicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/departments.DepartmentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get department by public ID
      tags:
      - departments
    put:
      consumes:
      - application/json
      description: Updates an existing department
      parameters:
      - description: Department Public ID
        in: path
        name: publicId
        required: true
        type: string
      - description: Department data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/departments.UpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/departments.DepartmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update department
      tags:
      - departments
  /departments/{publicId}/children:
    get:
      consumes:
      - application/json
      description: Retrieves direct children of a department
      parameters:
      - description: Department Public ID
        in: path
        name: publicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/departments.DepartmentResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get department children
      tags:
      - departments
  /departments/{publicId}/move:
    put:
      consumes:
      - application/json
      description: Moves a department to a new parent
      parameters:
      - description: Department Public ID
        in: path
        name: publicId
        required: true
        type: string
      - description: Move request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/departments.MoveDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/departments.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Move department
      tags:
      - departments
  /departments/{publicId}/tree:
    get:
      consumes:
      - application/json
      description: Retrieves a subtree starting from a specific department
      parameters:
      - description: Department Public ID
        in: path
        name: publicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/departments.DepartmentTreeResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get department subtree
      tags:
      - departments
  /departments/batch:
    delete:
      consumes:
      - application/json
      description: Soft-deletes multiple departments in a single transaction
      parameters:
      - description: Batch delete request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/departments.BatchDeleteDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/departments.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch delete departments
      tags:
      - departments
    post:
      consumes:
      - application/json
      description: Creates multiple departments in a single transaction
      parameters:
      - description: Batch create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/departments.BatchCreateDepartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/departments.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch create departments
      tags:
      - departments
    put:
      consumes:
      - application/json
      description: Updates multiple departments in a single transaction
      parameters:
      - description: Batch update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/departments.BatchUpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/departments.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch update departments
      tags:
      - departments
  /departments/search:
    post:
      consumes:
      - application/json
      description: Searches for departments based on name or description
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search criteria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/departments.SearchDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/departments.DepartmentListResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search departments
      tags:
      - departments
  /departments/tree:
    get:
      consumes:
      - application/json
      description: Retrieves the full department tree structure
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/departments.DepartmentTreeResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/departments.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get department tree
      tags:
      - departments
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"